requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
dataclasses-json==0.6.7
distro==1.9.0
docstring_parser==0.16
fastapi==0.115.11
firecrawl-py==1.14.1
frozenlist==1.5.0
gitdb==4.0.12
GitPython==3.1.44
google_search_results==2.4.2
greenlet==3.1.1
groq==0.19.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.20
langchain-community==0.3.19
langchain-core==0.3.45
langchain-groq==0.2.5
langchain-text-splitters==0.3.6
langsmith==0.3.15
markdown-it-py==3.0.0
marshmallow==3.26.1
mdurl==0.1.2
multidict==6.2.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numpy==2.2.4
orjson==3.10.15
packaging==24.2
phidata==2.7.10
propcache==0.3.0
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
Pygments==2.19.1
python-dotenv==1.0.1
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
serpapi==0.1.5
shellingham==1.5.4
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.39
starlette==0.46.1
tenacity==9.0.0
tomli==2.2.1
typer==0.15.2
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
websockets==15.0.1
yarl==1.18.3
zstandard==0.23.0
